Migrate remaining tests and type cleanup
