+++ browser_tests/tests/dialog.spec.ts
@@ -2,6 +2,7 @@ import { expect } from '@playwright/test'
 
 import type { Keybinding } from '@/platform/keybindings/types'
 import { comfyPageFixture as test } from '@e2e/fixtures/ComfyPage'
+import { SignInDialog } from '@e2e/fixtures/components/SignInDialog'
 import { DefaultGraphPositions } from '@e2e/fixtures/constants/defaultGraphPositions'
 
 test.beforeEach(async ({ comfyPage }) => {
@@ -151,70 +152,206 @@ test.describe('Signin dialog', () => {
 
     await expect.poll(() => comfyPage.nodeOps.getNodeCount()).toBe(nodeNum)
   })
+  //FIXME: neither of these tests are useful both need rewrites
+  test('Sign-in dialog resolves false when closed without sign-in', async ({
+    comfyPage
+  }) => {
+    type AugmentedWindow = Window &
+      typeof globalThis & { signInPromise?: Promise<boolean> }
+    await comfyPage.page.evaluate(async () => {
+      ;(window as AugmentedWindow).signInPromise =
+        window.app!.extensionManager.dialog.showSignInDialog()
+    })
+
+    const dialog = new SignInDialog(comfyPage.page)
+    await dialog.waitForVisible()
+
+    await dialog.close()
+    await expect(dialog.root).not.toBeVisible()
+
+    expect(
+      await comfyPage.page.evaluate(
+        () => (window as AugmentedWindow).signInPromise
+      )
+    ).toBe(false)
+  })
+
+  test('Sign-in dialog shows sign-in form and can fill credentials', async ({
+    comfyPage
+  }) => {
+    await comfyPage.page.evaluate(() => {
+      void window.app!.extensionManager.dialog.showSignInDialog()
+    })
+
+    const dialog = new SignInDialog(comfyPage.page)
+    await dialog.waitForVisible()
+
+    await dialog.emailInput.fill('test@example.com')
+    await expect(dialog.emailInput).toHaveValue('test@example.com')
+
+    await dialog.passwordInput.fill('TestPassword123!')
+    await expect(dialog.passwordInput).toHaveValue('TestPassword123!')
+
+    await expect(dialog.signInButton).toBeEnabled()
+  })
 })
 
-test(
-  'Blueprint overwrite',
-  { tag: ['@ui', '@subgraph'] },
-  async ({ comfyPage }) => {
-    await comfyPage.settings.setSetting('Comfy.VueNodes.Enabled', true)
-    const blueprintName = `test-blueprint-overwrite-${Date.now()}`
-    await comfyPage.settings.setSetting(
-      'Comfy.Workflow.WarnBlueprintOverwrite',
-      true
-    )
-
-    const tab = comfyPage.menu.nodeLibraryTabV2
-    await test.step('Publish a basic subgraph', async () => {
-      const ksampler = await comfyPage.vueNodes.getFixtureByTitle('KSampler')
-      await comfyPage.contextMenu.openForVueNode(ksampler.header)
-      await comfyPage.contextMenu.clickMenuItemExact('Convert to Subgraph')
-      await comfyPage.subgraph.publishSubgraph(blueprintName)
-
-      await tab.open()
-      await tab.getFolder('My Blueprints').click()
-      await tab.getFolder('User').click()
+test.describe('API Nodes sign-in dialog', () => {
+  test('Should display dialog with node names and resolve false on cancel', async ({
+    comfyPage
+  }) => {
+    await comfyPage.page.evaluate(() => {
+      void window
+        .app!.extensionManager.dialog.showApiNodesSignInDialog([
+          'FluxProGenerate',
+          'StableDiffusion3Generate'
+        ])
+        .then((result: boolean) => {
+          ;(window as unknown as Record<string, unknown>)['apiSignInResult'] =
+            result
+        })
     })
 
-    const steps = comfyPage.vueNodes.getWidgetByName('KSampler', 'steps')
-    await test.step('Edit the published subgraph', async () => {
-      const blueprintNode = tab.getNode(blueprintName)
-      await expect(blueprintNode, 'blueprint visible in library').toBeVisible()
-      await blueprintNode.getByRole('button', { name: 'Edit' }).click()
-      await steps.waitFor({ state: 'visible' })
+    const dialog = comfyPage.page.locator('.p-dialog')
+    await expect(dialog).toBeVisible()
+
+    await expect(
+      dialog.getByText('Sign In Required to Use API Nodes')
+    ).toBeVisible()
+    await expect(dialog.getByText('FluxProGenerate')).toBeVisible()
+    await expect(dialog.getByText('StableDiffusion3Generate')).toBeVisible()
+
+    await dialog.getByRole('button', { name: 'Cancel' }).click()
+    await expect(dialog).not.toBeVisible()
+
+    expect(
+      await comfyPage.page.evaluate(
+        () => (window as unknown as Record<string, unknown>)['apiSignInResult']
+      )
+    ).toBe(false)
+  })
+})
+
+test.describe('Update password dialog', () => {
+  test('Should open update password dialog with form fields', async ({
+    comfyPage
+  }) => {
+    await comfyPage.page.evaluate(() => {
+      void window.app!.extensionManager.dialog.showUpdatePasswordDialog()
     })
 
-    const confirmDialog = comfyPage.confirmDialog.root
-    const { incrementButton } = comfyPage.vueNodes.getInputNumberControls(steps)
-    const dirtyGraphAndSave = async () => {
-      await incrementButton.click()
-      await comfyPage.page.keyboard.press('Control+s')
-    }
+    const dialog = comfyPage.page.locator('.p-dialog')
+    await expect(dialog).toBeVisible()
 
-    await test.step('No dialog: user prompted on publish', async () => {
-      await dirtyGraphAndSave()
-      await comfyPage.nextFrame()
-      await expect(confirmDialog).not.toBeVisible()
+    await expect(dialog.locator('#comfy-org-sign-up-password')).toBeVisible()
+    await expect(
+      dialog.locator('#comfy-org-sign-up-confirm-password')
+    ).toBeVisible()
+    await expect(
+      dialog.getByRole('button', { name: 'Update Password' })
+    ).toBeVisible()
+  })
+})
+
+test.describe('Cloud notification dialog', () => {
+  test('Should display cloud notification and navigate to comfy.org on Explore', async ({
+    comfyPage
+  }) => {
+    await comfyPage.page.evaluate(() => {
+      void window.app!.extensionManager.dialog.showCloudNotification()
     })
 
-    await test.step('Should show dialog', async () => {
-      await comfyPage.subgraph.setSaveUnpromptedOnActiveBlueprint()
-      await dirtyGraphAndSave()
-      const { noWarnOverwriteToggle } = comfyPage.confirmDialog
-      await expect(noWarnOverwriteToggle).toBeVisible()
+    const dialog = comfyPage.page.locator('.p-dialog')
+    await expect(dialog).toBeVisible()
 
-      await test.step('Disable overwrite warning', async () => {
-        await noWarnOverwriteToggle.check()
-        await expect(confirmDialog.getByText(/Re-enable in/i)).toBeVisible()
-        await comfyPage.confirmDialog.click('overwrite')
-      })
+    await expect(dialog.getByText('Run ComfyUI in the Cloud')).toBeVisible()
+    await expect(
+      dialog.getByRole('button', { name: 'Continue Locally' })
+    ).toBeVisible()
+
+    const popupPromise = comfyPage.page.waitForEvent('popup')
+    await dialog.getByRole('button', { name: 'Try Cloud for Free' }).click()
+    const popup = await popupPromise
+
+    expect(new URL(popup.url()).hostname).toContain('comfy.org')
+    await popup.close()
+    await expect(dialog).not.toBeVisible()
+  })
+
+  test('Should close when Continue Locally is clicked', async ({
+    comfyPage
+  }) => {
+    await comfyPage.page.evaluate(() => {
+      void window.app!.extensionManager.dialog.showCloudNotification()
     })
 
-    await test.step('No dialog: disabled by setting', async () => {
-      await comfyPage.subgraph.setSaveUnpromptedOnActiveBlueprint()
-      await dirtyGraphAndSave()
-      await comfyPage.nextFrame()
-      await expect(confirmDialog).not.toBeVisible()
-    })
+    const dialog = comfyPage.page.locator('.p-dialog')
+    await expect(dialog).toBeVisible()
+
+    await dialog.getByRole('button', { name: 'Continue Locally' }).click()
+    await expect(dialog).not.toBeVisible()
+  })
+})
+
+test('Blueprint overwrite', { tag: ['@subgraph'] }, async ({ comfyPage }) => {
+  const blueprintName = `test-blueprint-overwrite-${Date.now()}`
+  await comfyPage.settings.setSetting('Comfy.UseNewMenu', 'Top')
+  await comfyPage.settings.setSetting('Comfy.VueNodes.Enabled', true)
+  await comfyPage.settings.setSetting(
+    'Comfy.Workflow.WarnBlueprintOverwrite',
+    true
+  )
+
+  const tab = comfyPage.menu.nodeLibraryTabV2
+  await test.step('Publish a basic subgraph', async () => {
+    const ksampler = await comfyPage.vueNodes.getFixtureByTitle('KSampler')
+    await comfyPage.contextMenu.openForVueNode(ksampler.header)
+    await comfyPage.contextMenu.clickMenuItemExact('Convert to Subgraph')
+    await comfyPage.subgraph.publishSubgraph(blueprintName)
+
+    await tab.open()
+    await tab.getFolder('My Blueprints').click()
+    await tab.getFolder('User').click()
+  })
+
+  const steps = comfyPage.vueNodes.getWidgetByName('KSampler', 'steps')
+  await test.step('Edit the published subgraph', async () => {
+    const blueprintNode = tab.getNode(blueprintName)
+    await expect(blueprintNode, 'blueprint visible in library').toBeVisible()
+    await blueprintNode.getByRole('button', { name: 'Edit' }).click()
+    await steps.waitFor({ state: 'visible' })
+  })
+
+  const confirmDialog = comfyPage.confirmDialog.root
+  const { incrementButton } = comfyPage.vueNodes.getInputNumberControls(steps)
+  const dirtyGraphAndSave = async () => {
+    await incrementButton.click()
+    await comfyPage.page.keyboard.press('Control+s')
   }
-)
+
+  await test.step('No dialog: user prompted on publish', async () => {
+    await dirtyGraphAndSave()
+    await comfyPage.nextFrame()
+    await expect(confirmDialog).not.toBeVisible()
+  })
+
+  await test.step('Should show dialog', async () => {
+    await comfyPage.subgraph.setSaveUnpromptedOnActiveBlueprint()
+    await dirtyGraphAndSave()
+    const { noWarnOverwriteToggle } = comfyPage.confirmDialog
+    await expect(noWarnOverwriteToggle).toBeVisible()
+
+    await test.step('Disable overwrite warning', async () => {
+      await noWarnOverwriteToggle.check()
+      await expect(confirmDialog.getByText(/Re-enable in/i)).toBeVisible()
+      await comfyPage.confirmDialog.click('overwrite')
+    })
+  })
+
+  await test.step('No dialog: disabled by setting', async () => {
+    await comfyPage.subgraph.setSaveUnpromptedOnActiveBlueprint()
+    await dirtyGraphAndSave()
+    await comfyPage.nextFrame()
+    await expect(confirmDialog).not.toBeVisible()
+  })
+})
