+++ browser_tests/tests/errorDialog.spec.ts
@@ -137,4 +137,57 @@ test.describe('Error dialog', () => {
 
     await popup.close()
   })
+
+  test('Should display extension file hint when available', async ({
+    comfyPage
+  }) => {
+    await comfyPage.page.evaluate(() => {
+      const error = new Error('Extension error!')
+      ;(error as Error & { fileName: string }).fileName =
+        '/extensions/my-custom-extension/main.js'
+
+      window.app!.extensionManager.dialog.showErrorDialog(error)
+    })
+
+    const errorDialog = comfyPage.page.locator('.comfy-error-report')
+    await expect(errorDialog).toBeVisible()
+
+    await expect(
+      errorDialog.getByText('/extensions/my-custom-extension/main.js')
+    ).toBeVisible()
+    await expect(
+      errorDialog.getByText('This may be due to the following script')
+    ).toBeVisible()
+  })
+  test('Should display string error messages', async ({ comfyPage }) => {
+    await comfyPage.page.evaluate(() => {
+      window.app!.extensionManager.dialog.showErrorDialog(
+        'Something went wrong',
+        {
+          title: 'Custom Error Title'
+        }
+      )
+    })
+
+    const errorDialog = comfyPage.page.locator('.comfy-error-report')
+    await expect(errorDialog).toBeVisible()
+
+    await expect(errorDialog.getByText('Custom Error Title')).toBeVisible()
+    await expect(errorDialog.getByText('Something went wrong')).toBeVisible()
+  })
+  test('Should display default title when no title provided', async ({
+    comfyPage
+  }) => {
+    await comfyPage.page.evaluate(() => {
+      window.app!.extensionManager.dialog.showErrorDialog(
+        'A simple string error'
+      )
+    })
+
+    const errorDialog = comfyPage.page.locator('.comfy-error-report')
+    await expect(errorDialog).toBeVisible()
+
+    await expect(errorDialog.getByText('Unknown Error')).toBeVisible()
+    await expect(errorDialog.getByText('A simple string error')).toBeVisible()
+  })
 })
